Git Ignore update
This now ignores any folder with the name Bin, bin, Binary, or binary.
Which means the DKTK build files, and root Binart build files are no
longer part of the project.
Also, for some reason sometimes git translates ".sdf" to ".opensdf" so
that's been added as well, cause we don't want sql crap.