Memory Leak stuff
Added a line in BoatGame.cpp's destructor that will print all the memory
leaks at the end of the application. (If this line throws an error for
anyone let me know and we can take it out and re-commit.)

We need to fix the memory leaks that are back in the solution.