Engine resource fix
So, engine seemed to have is own resource compiler with a bad direct
path and no resource.h to boot.
Deleting it made things compile, run, and the icon still seems to come
up on both windows.
Also, deleting a .aps file, this is a compiled .rc file which i will add
to the gitIgnore after this commit (otherwise it will ignore my delete
too!)