Entity Work
There is now a "world" Entity in Gameplay that will act as the parent
for all entities that would not already have parents (Boat, Water,
etc.). Now, because the world entity is the parent to anything only it
needs to be Updated and everything will "trickle down". Thus, destroying
the world entity will destroy all entities (i.e. all its children).

Vector of Entities is now a vector of DrawableEntities, used only to
reference those entities which need to be drawn (since Entity does not
have Draw()).

Entity can now base its orientation on EITHER a vector of rotation
angles or a forward vector - this is decided in the constructor for
Entity (default parameter is that it will rely on rotation angles). This
could be vastly improved.

Added roll to Entity for a 3rd-dimension rotation.

Moved boat's "death animation" our of Gameplay and into the boat class.