Rotating Camera
The camera now rotates around the boat (fixed center of the boat).
Movement of the camera is a bit janky and will be worked out, but it is
working. Need InputManager to include some stuff like hiding the mouse,
locking the mouse to the center of the screen, taking into account when
the mouse goes off screen, etc.

Resources can now be unloaded with an Unload() method in GameState
(rather than in DX Release()) which releases all resources when the
GameState is changed.