Water Simulation! (Nearly complete)
Water now moves more realistically (and more correctly in the math).

Duplicated math from the shader to find where the boats *center* should
rest on the water. Soon, the boat will rotate according to exactly how
it should rest on the water.

Game forward declaration in Boat class. This will most likely be
removed, but it was the easiest way to do this and does not hurt/break
anything else.